move ActivityRoute to its own file
