+++ navigation/api/navigation.api
@@ -5,6 +5,11 @@ public abstract interface class com/freeletics/khonshu/navigation/ActivityRoute 
 	public abstract fun fillInIntent ()Landroid/content/Intent;
 }
 
+public final class com/freeletics/khonshu/navigation/ActivityRouteKt {
+	public static final fun getRoute (Landroid/app/Activity;)Lcom/freeletics/khonshu/navigation/InternalActivityRoute;
+	public static final fun requireRoute (Landroid/app/Activity;)Lcom/freeletics/khonshu/navigation/InternalActivityRoute;
+}
+
 public abstract interface class com/freeletics/khonshu/navigation/BaseRoute : android/os/Parcelable {
 }
 
@@ -102,11 +107,6 @@ public abstract interface class com/freeletics/khonshu/navigation/NavRoot : com/
 }
 
 public abstract interface class com/freeletics/khonshu/navigation/NavRoute : com/freeletics/khonshu/navigation/BaseRoute {
-}
-
-public final class com/freeletics/khonshu/navigation/NavRouteKt {
-	public static final fun getRoute (Landroid/app/Activity;)Lcom/freeletics/khonshu/navigation/InternalActivityRoute;
-	public static final fun requireRoute (Landroid/app/Activity;)Lcom/freeletics/khonshu/navigation/InternalActivityRoute;
 }
 
 public final class com/freeletics/khonshu/navigation/NavigationResultRequest : com/freeletics/khonshu/navigation/ResultOwner {
